Creato nuovo file di preprocessing, rimosse le colonne
Con questo commit è finita definitivamente la fase di preprocessing (salvo imprevisti). Ho finito di sistemare il nuovo file 'preprocessing.py' e dovrebbe essere tutto apposto. @EC2002 quando vuoi fai la review e fai merge;
